cleanup and move api schema
