+++ src/main/java/io/github/dqualizer/dqlang/types/rqa/definition/RuntimeQualityAnalysis.java
@@ -2,8 +2,8 @@ package io.github.dqualizer.dqlang.types.rqa.definition;
 
 
 import io.github.dqualizer.dqlang.types.rqa.definition.loadtest.LoadTestDefinition;
-import io.github.dqualizer.dqlang.types.rqa.def.monitoring.MonitoringDefinition;
-import io.github.dqualizer.dqlang.types.rqa.def.resilience.ResilienceDefinition;
+import io.github.dqualizer.dqlang.types.rqa.definition.monitoring.MonitoringDefinition;
+import io.github.dqualizer.dqlang.types.rqa.definition.resilience.ResilienceDefinition;
 import lombok.Data;
 
 import java.util.HashSet;
