+++ src/main/kotlin/io/github/dqualizer/dqlang/types/dam/architecture/apischema/data/DataSchema.java
@@ -1,16 +1,18 @@
-package io.github.dqualizer.dqlang.types.apiadapter.data;
+package io.github.dqualizer.dqlang.types.dam.architecture.apischema.data;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+import java.util.Set;
+
 @Getter
 @ToString
 @AllArgsConstructor
 @NoArgsConstructor
-public class Property {
+public class DataSchema {
 
-    private String name;
     private String type;
+    private Set<Property> properties;
 }
