+++ src/main/kotlin/io/github/dqualizer/dqlang/types/dam/architecture/apischema/field/DataType.java
@@ -1,4 +1,4 @@
-package io.github.dqualizer.dqlang.types.apiadapter.field;
+package io.github.dqualizer.dqlang.types.dam.architecture.apischema.field;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
