--- src/main/kotlin/io/github/dqualizer/dqlang/types/rqa/def/resilience/ResilienceDefinition.kt
@@ -1,3 +1,0 @@
-package io.github.dqualizer.dqlang.types.rqa.def.resilience
-
-class ResilienceDefinition
