+++ src/main/kotlin/io/github/dqualizer/dqlang/types/rqa/definition/resilience/ResilienceDefinition.kt
@@ -1,0 +1,3 @@
+package io.github.dqualizer.dqlang.types.rqa.definition.resilience
+
+class ResilienceDefinition
